+++ src/main/java/com/agenticcp/core/domain/cloud/adapter/outbound/AccountCredentialPortRouter.java
@@ -9,9 +9,9 @@ import com.agenticcp.core.domain.cloud.port.model.CloudSessionCredential;
 import com.agenticcp.core.domain.cloud.port.model.account.AccountValidationRequest;
 import com.agenticcp.core.domain.cloud.port.model.account.AccountValidationResult;
 import com.agenticcp.core.domain.cloud.port.model.account.ConnectionTestResult;
-import com.agenticcp.core.domain.cloud.port.outbound.AccountCredentialManagementPort;
-import com.agenticcp.core.domain.cloud.port.outbound.AccountSyncPort;
-import com.agenticcp.core.domain.cloud.port.outbound.AccountValidationPort;
+import com.agenticcp.core.domain.cloud.port.outbound.account.AccountCredentialManagementPort;
+import com.agenticcp.core.domain.cloud.port.outbound.account.AccountSyncPort;
+import com.agenticcp.core.domain.cloud.port.outbound.account.AccountValidationPort;
 import com.agenticcp.core.domain.cloud.repository.CloudAccountRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Primary;
