+++ src/main/java/com/agenticcp/core/domain/cloud/adapter/outbound/aws/account/AwsAccountCredentialManagementAdapter.java
@@ -6,7 +6,7 @@ import com.agenticcp.core.domain.cloud.entity.CloudProvider;
 import com.agenticcp.core.domain.cloud.entity.CloudProvider.ProviderType;
 import com.agenticcp.core.domain.cloud.exception.CloudErrorCode;
 import com.agenticcp.core.domain.cloud.port.model.CloudSessionCredential;
-import com.agenticcp.core.domain.cloud.port.outbound.AccountCredentialManagementPort;
+import com.agenticcp.core.domain.cloud.port.outbound.account.AccountCredentialManagementPort;
 import com.agenticcp.core.domain.cloud.repository.CloudAccountRepository;
 import com.agenticcp.core.domain.cloud.service.SessionCacheService;
 import lombok.RequiredArgsConstructor;
