+++ src/main/java/com/agenticcp/core/domain/cloud/config/CloudStubConfiguration.java
@@ -3,7 +3,7 @@ package com.agenticcp.core.domain.cloud.config;
 import com.agenticcp.core.domain.cloud.entity.CloudProvider.ProviderType;
 import com.agenticcp.core.domain.cloud.port.model.CloudSessionCredential;
 import com.agenticcp.core.domain.cloud.port.outbound.AuditEventPort;
-import com.agenticcp.core.domain.cloud.port.outbound.CredentialProviderPort;
+import com.agenticcp.core.domain.cloud.port.outbound.account.AccountCredentialManagementPort;
 import com.agenticcp.core.domain.cloud.port.outbound.OutboxEventPort;
 import com.agenticcp.core.domain.cloud.port.outbound.TracingPort;
 import lombok.extern.slf4j.Slf4j;
@@ -30,9 +30,9 @@ public class CloudStubConfiguration {
 
     @Bean
     @Primary
-    public CredentialProviderPort credentialProviderPort() {
-        log.info("Cloud Stub implementation enabled for CredentialProviderPort");
-        return new CredentialProviderPort() {
+    public AccountCredentialManagementPort accountCredentialManagementPort() {
+        log.info("Cloud Stub implementation enabled for AccountCredentialManagementPort");
+        return new AccountCredentialManagementPort() {
             @Override
             public Object resolveCredentials(String tenantKey, ProviderType providerType, String accountScope) {
                 log.debug("Stub: Credential resolution - tenant={}, provider={}, account={}",
